adds tests for some methods of the job executor

+++ service/processor.go
@@ -158,7 +158,7 @@ func (jobProcessor *JobProcessor) RecoverJobs() {
 			// comparing the execution time on the schedule by recalculating it against it's time.
 			// Time clocks are sources of distributed systems errors and a monotonic clock should always be preferred.
 			// While 60 minutes is quite an unlike delay in a close it's not impossible
-			if now.Before(executionTime) && lastJobState.State == uint64(models.ExecutionLogScheduleState) {
+			if now.Before(executionTime) && lastJobState.State == models.ExecutionLogScheduleState {
 				jobProcessor.logger.Debug(fmt.Sprintf("quick recovered job %d", job.ID))
 				jobProcessor.jobExecutor.ScheduleProcess(job, executionTime)
 			} else {
